Add Consolidate Centerlines Tool
Tested on a small subset of OSM data ,this tool will consolidate
centerlines and get the mean, sum, first, and concatenated values of
selected collapsed fields. For lanes you can get the sum for example,
speed limits the mean, names the first value, and just in case the
ability to concat all values. Needs documentation in toolbox, but this
draft works.