Add Possible Value Column to Additive Spec
Possible values column added to Additive Spec. As we work on this, I am
starting to wonder if we should use coded values (Enums) instead of
text. If we add Domains to a database, it would be human readable
regardless. @louh  or @marsxul  any thoughts on this? I know GTFS does
this for some fields.