Minimum Viable Crosswalk Script
This has taken a while to develop, but this script provides a minimum
viable converter between the additive specification and the slice based
specification. A convert to a geojson from this crosswalk is the next
item to work on. This will need the following improvements:
1. Better conditional treatment of metatags- none is provided now.
2. Potentially breaking this out into tables and reference geometry, or
just outputing a table with a join ID to the input geometry. The
replication of geometry is tedious. Another option is only the first
record of every street has geometry in the feature class, the the rest
are blank.

@marsxul any thoughts on the second point?

Commit includes tool, test database, and python code edits.