Change Turn Lane Defaults
Change turn lanes to default left, and only be forward if they are short
segments. @louh  - I know we discussed transitions, and as I work on
this, do you think there should be flag to indicate whether a segment is
block to block, mid-block, or includes an approach? Might not be too
important for the default converters...not sure .