Add Slice Exporter Draft
Remove test json, make a slice spec sample. @louh  and @marsxul - can
you both review the sample slice_spec_sample.geojson? My concern is the
escape characters around the export in the slices. After this I might
move on to CityEngine export examples. @louh do you want to make any
samples for streetmix reading in the data? It could be as basic as a
leaflet map, and a resulting street in streetmix based on someone
clicking on a segment (then again that might not be basic!).
We should iron out this geojson exporter first though.